Shift `loginUser` function into `@/services/users`
